Adjust Startup Behavior (#17)
React to arguments provided via command line and adjust the startup behavior accordingly.

When a directory is provided as argument, it will be opened.

When one or more arguments are provided, they are treated as files and will be decoded. Pressing ESC will cause the current dir to be opened.