Fix a deadlock when calling SmartImageDecoder::cancelOrTake() twice
This is because the first call might surprisingly make isFinished and taken ==true, causing the future to never correctly end up in finished state. Therefore, the second call will not recognized that the future has already finished, causing it to cancel the future again, making it enter the canceled state (but NOT finished state), causing any later call to waitForFinished() to wait forever.