Switch to new Windows CI container (#53)
The new Windows CI container uses QT 6.5.3. This is a necessary preparation for JXL support.

Unfortunately, this breaks the build because:

1. QByteArray::isNull is now multiply defined in many compilation units, and
2. Transitive dependencies of Exiv2 are not taken care of by KDE's CMake modules.

2) will be fixed later.

The reason for 1) is that libkexiv2 and libkdcraw are using different deprecation levels than ANPV. When linked together, we will get multiple definitions of the same function, see the last comment in this issue:
https://bugreports.qt.io/browse/QTBUG-109449
Solution is to simply remove the deprecation levels for libkdcraw and likexiv2.