Fix kdcraw being unable to open some RAW files
kdcraw might be unable to open a RAW file if it is located in a folder which has special characters. Might be an encoding issue somewhere. Only Release builds on Windows were affected. Get rid of all that filepath conversion stuff and use LibRaw directly to open the existing file buffer.